*Minor changes to satisfly Cppcheck and Splint.  The loops in eof_gh_read_sections_qb() were modified so that the return statement at the end of the function is actually used. *Fixed a bug where when a lane 5 drum gem was combined with any other gems, those other gems were automatically removed in the way that open strum guitar notes remove other gems. *Fixed a bug in the new arpeggio cleanup logic that could cause EOF to hang.

+++ src/foflc/RS_parse.c
@@ -60,7 +60,7 @@ void Export_RS(FILE *outf)
 		while(temp != NULL)				//For each piece of lyric in this line
 		{
 			expand_xml_text(buffer2, sizeof(buffer2) - 1, temp->lyric, 32);	//Expand XML special characters into escaped sequences if necessary, and check against the maximum supported length of this field
-			for(index1 = index2 = 0; index1 < strlen(buffer2); index1++)
+			for(index1 = index2 = 0; (size_t)index1 < strlen(buffer2); index1++)
 			{	//For each character in the expanded XML string
 				if(buffer2[index1] != '+')
 				{	//If it's not a plus character
